Fix MacOS compilation

+++ src/pgduckdb_types.cpp
@@ -446,7 +446,8 @@ DatumGetTimeTz(Datum value) {
 	TimeTzADT *tzt = static_cast<TimeTzADT *>(DatumGetTimeTzADTP(value));
 	// pg and duckdb stores timezone with different signs, for example, for TIMETZ 01:02:03+05, duckdb stores offset =
 	// 18000, while pg stores zone = -18000.
-	const uint64_t bits = duckdb::dtime_tz_t::encode_micros(tzt->time) | duckdb::dtime_tz_t::encode_offset(-tzt->zone);
+	const uint64_t bits = duckdb::dtime_tz_t::encode_micros(static_cast<int64_t>(tzt->time)) |
+	                      duckdb::dtime_tz_t::encode_offset(-tzt->zone);
 	const duckdb::dtime_tz_t duck_time_tz {bits};
 	return duck_time_tz;
 }
